// old/ferdikloomen/gen/models/JsonRpcRequest_for_BlockMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BlockMethodNameHelperEnum } from './BlockMethodNameHelperEnum';
import type { RpcBlockRequest } from './RpcBlockRequest';
export type JsonRpcRequest_for_BlockMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: BlockMethodNameHelperEnum;
    params: RpcBlockRequest;
};