// old/ferdikloomen/gen/models/JsonRpcRequest_for_BroadCastTxAsyncMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BroadCastTxAsyncMethodNameHelperEnum } from './BroadCastTxAsyncMethodNameHelperEnum';
import type { RpcSendTransactionRequest } from './RpcSendTransactionRequest';
export type JsonRpcRequest_for_BroadCastTxAsyncMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: BroadCastTxAsyncMethodNameHelperEnum;
    params: RpcSendTransactionRequest;
};