// old/ferdikloomen/gen/models/JsonRpcRequest_for_BroadCastTxCommitMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BroadCastTxCommitMethodNameHelperEnum } from './BroadCastTxCommitMethodNameHelperEnum';
import type { RpcSendTransactionRequest } from './RpcSendTransactionRequest';
export type JsonRpcRequest_for_BroadCastTxCommitMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: BroadCastTxCommitMethodNameHelperEnum;
    params: RpcSendTransactionRequest;
};