// old/ferdikloomen/gen/models/JsonRpcRequest_for_ChunkMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ChunkMethodNameHelperEnum } from './ChunkMethodNameHelperEnum';
import type { RpcChunkRequest } from './RpcChunkRequest';
export type JsonRpcRequest_for_ChunkMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ChunkMethodNameHelperEnum;
    params: RpcChunkRequest;
};