// old/ferdikloomen/gen/models/JsonRpcRequest_for_ClientConfigMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ClientConfigMethodNameHelperEnum } from './ClientConfigMethodNameHelperEnum';
import type { RpcClientConfigRequest } from './RpcClientConfigRequest';
export type JsonRpcRequest_for_ClientConfigMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ClientConfigMethodNameHelperEnum;
    params: RpcClientConfigRequest;
};