// old/ferdikloomen/gen/models/JsonRpcRequest_for_ExpChangeMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ExpChangeMethodNameHelperEnum } from './ExpChangeMethodNameHelperEnum';
import type { RpcStateChangesInBlockByTypeRequest } from './RpcStateChangesInBlockByTypeRequest';
export type JsonRpcRequest_for_ExpChangeMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ExpChangeMethodNameHelperEnum;
    params: RpcStateChangesInBlockByTypeRequest;
};