// old/ferdikloomen/gen/models/JsonRpcRequest_for_ExpChangesBlockMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ExpChangesBlockMethodNameHelperEnum } from './ExpChangesBlockMethodNameHelperEnum';
import type { RpcStateChangesInBlockRequest } from './RpcStateChangesInBlockRequest';
export type JsonRpcRequest_for_ExpChangesBlockMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ExpChangesBlockMethodNameHelperEnum;
    params: RpcStateChangesInBlockRequest;
};