// old/ferdikloomen/gen/models/JsonRpcRequest_for_ExpGenesisMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ExpGenesisMethodNameHelperEnum } from './ExpGenesisMethodNameHelperEnum';
import type { GenesisConfigRequest } from './GenesisConfigRequest';
export type JsonRpcRequest_for_ExpGenesisMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ExpGenesisMethodNameHelperEnum;
    params: GenesisConfigRequest;
};