// old/ferdikloomen/gen/models/JsonRpcRequest_for_ExpGongestionMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ExpGongestionMethodNameHelperEnum } from './ExpGongestionMethodNameHelperEnum';
import type { RpcCongestionLevelRequest } from './RpcCongestionLevelRequest';
export type JsonRpcRequest_for_ExpGongestionMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ExpGongestionMethodNameHelperEnum;
    params: RpcCongestionLevelRequest;
};