// old/ferdikloomen/gen/models/JsonRpcRequest_for_ExpLightClientBlockProofMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ExpLightClientBlockProofMethodNameHelperEnum } from './ExpLightClientBlockProofMethodNameHelperEnum';
import type { RpcLightClientBlockProofRequest } from './RpcLightClientBlockProofRequest';
export type JsonRpcRequest_for_ExpLightClientBlockProofMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ExpLightClientBlockProofMethodNameHelperEnum;
    params: RpcLightClientBlockProofRequest;
};