// old/ferdikloomen/gen/models/JsonRpcRequest_for_ExpLightClientProofMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ExpLightClientProofMethodNameHelperEnum } from './ExpLightClientProofMethodNameHelperEnum';
import type { RpcLightClientExecutionProofRequest } from './RpcLightClientExecutionProofRequest';
export type JsonRpcRequest_for_ExpLightClientProofMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ExpLightClientProofMethodNameHelperEnum;
    params: RpcLightClientExecutionProofRequest;
};