// old/ferdikloomen/gen/models/JsonRpcRequest_for_ExpMaintenanceWindoesMethodNameHelperEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ExpMaintenanceWindoesMethodNameHelperEnum } from './ExpMaintenanceWindoesMethodNameHelperEnum';
import type { RpcMaintenanceWindowsRequest } from './RpcMaintenanceWindowsRequest';
export type JsonRpcRequest_for_ExpMaintenanceWindoesMethodNameHelperEnum = {
    id: string;
    jsonrpc: string;
    method: ExpMaintenanceWindoesMethodNameHelperEnum;
    params: RpcMaintenanceWindowsRequest;
};